Fixed the time logic and the server but the server and client is intermittent on build for some reason, when tested in visualstudio code also inermittent on build

+++ NeuralNetwork.cs
@@ -99,7 +99,7 @@ namespace Thyme_new
             return finalOutputs;
         }
 
-        public void Train(double[] input, byte[] target, double learningRate)
+        public void Train(double[] input, double[] target, double learningRate)
         {
             Console.WriteLine("beginning training");
             if (input.Length != inputSize || target.Length != outputSize)
